8271024: Implement macOS Metal Rendering Pipeline

+++ modules/javafx.graphics/src/main/java/com/sun/glass/ui/ios/IosView.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2011, 2020, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2011, 2025, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -60,7 +60,7 @@ final class IosView extends View {
     // See View
     @Override protected void _enableInputMethodEvents(long ptr, boolean enable) { }
 
-    @Override native protected int _getNativeFrameBuffer(long ptr);
+    @Override native protected long _getNativeFrameBuffer(long ptr);
     @Override native protected long _create(Map caps);
 
     @Override native protected long _getNativeView(long ptr);
